requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
polygon-api-client==1.13.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
apscheduler==3.10.4
pytz==2023.3

